[DEBT-043-D2] Split cli/data.py → cli/data/ package
- Split cli/data.py (737 LoC) into cli/data/ package with 10 files
- Largest file: sync.py at 225 LoC (well under 400 LoC ceiling)
- Files created: _helpers.py, init_cmd.py, migrate.py, sync.py,
  snapshot.py, collect.py, export_cmd.py, stats.py, maintenance.py,
  __init__.py
- All tests pass (1052 passed)
- Quality gates: pre-commit, mypy, pytest
- Acceptance criteria: 1/1 complete (Phase D2)
