+++ src/kalshi_research/cli/data/export_cmd.py
@@ -1,0 +1,49 @@
+"""Data export command."""
+
+from pathlib import Path
+from typing import Annotated
+
+import typer
+from rich.progress import Progress, SpinnerColumn, TextColumn
+
+from kalshi_research.cli.utils import console
+from kalshi_research.paths import DEFAULT_DB_PATH, DEFAULT_EXPORTS_DIR
+
+
+def data_export(
+    db_path: Annotated[
+        Path,
+        typer.Option("--db", "-d", help="Path to SQLite database file."),
+    ] = DEFAULT_DB_PATH,
+    output: Annotated[
+        Path,
+        typer.Option("--output", "-o", help="Output directory for exports."),
+    ] = DEFAULT_EXPORTS_DIR,
+    format_type: Annotated[
+        str,
+        typer.Option("--format", "-f", help="Export format (parquet, csv)."),
+    ] = "parquet",
+) -> None:
+    """Export data to Parquet or CSV for analysis."""
+    from kalshi_research.data.export import export_to_csv, export_to_parquet
+
+    if not db_path.exists():
+        console.print(f"[red]Error:[/red] Database not found at {db_path}")
+        raise typer.Exit(1)
+
+    with Progress(
+        SpinnerColumn(),
+        TextColumn("[progress.description]{task.description}"),
+        console=console,
+    ) as progress:
+        progress.add_task(f"Exporting to {format_type}...", total=None)
+
+        if format_type == "parquet":
+            export_to_parquet(db_path, output)
+        elif format_type == "csv":
+            export_to_csv(db_path, output)
+        else:
+            console.print(f"[red]Error:[/red] Unknown format: {format_type}")
+            raise typer.Exit(1)
+
+    console.print(f"[green]✓[/green] Exported to {output}")
